refactor(web-app): move IconActionButton to components folder
